+++ test/Arpeggio.test.ts
@@ -7,7 +7,7 @@ import { Tween } from '../src';
 
 @suite class ArpeggioTests {
     @test 'createNote returns null if no chord passed in'() {
-        const arpNote = new ArpeggioNote(0, 1, c => c.getPitch(0), 80);
+        const arpNote = new ArpeggioNote(0, 1, c => c.getPitchByIndex(0), 80);
         expect(arpNote.createNote(null, 0, 0)).to.be.null;
         expect(arpNote.createNote(undefined, 0, 0)).to.be.null;
     }
@@ -32,7 +32,7 @@ import { Tween } from '../src';
 
     @test 'createNote returns new note object'() {
         const chord = new Chord().addPitches([4, 8, 11]);
-        const arpNote = new ArpeggioNote(0, 1, c => c.getPitch(1), 81, 3);
+        const arpNote = new ArpeggioNote(0, 1, c => c.getPitchByIndex(1), 81, 3);
         const note = arpNote.createNote(chord, 1, 0);
         expect(note.pitch).to.equal(8);
         expect(note.velocity).to.equal(81);
@@ -41,7 +41,7 @@ import { Tween } from '../src';
 
     @test 'createNote takes percent for setting initial velocity tween value'() {
         const chord = new Chord().addPitches([4, 8, 11]);
-        const arpNote = new ArpeggioNote(0, 1, c => c.getPitch(1), Tween.linear(0, 80), 3);
+        const arpNote = new ArpeggioNote(0, 1, c => c.getPitchByIndex(1), Tween.linear(0, 80), 3);
         const note = arpNote.createNote(chord, 1, 0.5);
         expect(note.velocity).to.equal(40);
     }
@@ -49,10 +49,10 @@ import { Tween } from '../src';
     @test 'getNotesStartingInRange handles wrapping around start/end'() {
         const arp = new Arpeggio(4);
         arp.notes.push(
-            new ArpeggioNote(0, 1, c => c.getPitch(0), 80),
-            new ArpeggioNote(1, 1, c => c.getPitch(1), 81),
-            new ArpeggioNote(2, 1, c => c.getPitch(2), 82),
-            new ArpeggioNote(3, 1, c => c.getPitch(1), 83)
+            new ArpeggioNote(0, 1, c => c.getPitchByIndex(0), 80),
+            new ArpeggioNote(1, 1, c => c.getPitchByIndex(1), 81),
+            new ArpeggioNote(2, 1, c => c.getPitchByIndex(2), 82),
+            new ArpeggioNote(3, 1, c => c.getPitchByIndex(1), 83)
         );
         const startingNotes = arp.getNotesStartingInRange(3.9, 0.1);
 
@@ -63,10 +63,10 @@ import { Tween } from '../src';
     @test 'getNotesEndingInRange handles wrapping around start/end'() {
         const arp = new Arpeggio(4);
         arp.notes.push(
-            new ArpeggioNote(0, 1, c => c.getPitch(0), 80),
-            new ArpeggioNote(1, 1, c => c.getPitch(1), 81),
-            new ArpeggioNote(2, 1, c => c.getPitch(2), 82),
-            new ArpeggioNote(3, 1, c => c.getPitch(1), 83)
+            new ArpeggioNote(0, 1, c => c.getPitchByIndex(0), 80),
+            new ArpeggioNote(1, 1, c => c.getPitchByIndex(1), 81),
+            new ArpeggioNote(2, 1, c => c.getPitchByIndex(2), 82),
+            new ArpeggioNote(3, 1, c => c.getPitchByIndex(1), 83)
         );
         const endingNotes = arp.getNotesEndingInRange(3.9, 0.1);
         expect(endingNotes.length).to.equal(1);
@@ -75,7 +75,7 @@ import { Tween } from '../src';
 
     @test 'addNote can add new note'() {
         const arp = new Arpeggio(4);
-        arp.addNote(0, 1, c => c.getPitch(0), 80);
+        arp.addNote(0, 1, c => c.getPitchByIndex(0), 80);
         expect(arp.notes.length).to.equal(1);
         expect(arp.notes[0].start).to.equal(0);
         expect(arp.notes[0].duration).to.equal(1);
@@ -83,18 +83,18 @@ import { Tween } from '../src';
 
     @test 'addNote returns reference to parent arpeggio object'() {
         const arp = new Arpeggio(4)
-            .addNote(0, 1, c => c.getPitch(0), 80);
+            .addNote(0, 1, c => c.getPitchByIndex(0), 80);
         expect(arp.notes.length).to.equal(1);
     }
 
     @test 'addNote can take array of note starts'() {
         const arp = new Arpeggio(4)
-            .addNote([0,1,2,3], 1, c => c.getPitch(0), 80);
+            .addNote([0,1,2,3], 1, c => c.getPitchByIndex(0), 80);
         expect(arp.notes.length).to.equal(4);
     }
 
     @test 'addNote can take velocity tween'() {
         const arp = new Arpeggio(4)
-            .addNote(0, 1, c => c.getPitch(0), Tween.linear(1, 127));
+            .addNote(0, 1, c => c.getPitchByIndex(0), Tween.linear(1, 127));
     }
 }